requirements.txt
certifi==2021.10.8
geojson==2.5.0
idna==3.3
mapq==0.3
psycopg2==2.9.2
PySocks==1.7.1
ratelimit==2.2.1
requests==2.26.0
simplejson==3.17.6
tqdm==4.62.3
urllib3==1.26.7
pandas==1.3.5
scikit-learn==1.0.2
darkskylib==0.3.91
